tests/conftest.py
import pytest_asyncio
from httpx import AsyncClient, ASGITransport

from sqlalchemy.ext.asyncio import create_async_engine
from sqlmodel.ext.asyncio.session import AsyncSession
from sqlalchemy.orm import sessionmaker
from sqlmodel import SQLModel

from app.main import app
from app.core.database import get_session

# Use In-Memory SQLite for Tests to avoid polluting Real DB
TEST_DATABASE_URL = "sqlite+aiosqlite:///:memory:"

@pytest_asyncio.fixture(loop_scope="session", scope="session")
async def engine_fixture():
    engine = create_async_engine(
        TEST_DATABASE_URL, 
        echo=False, 
        future=True,
        connect_args={"check_same_thread": False} # Needed for SQLite
    )
    yield engine
    await engine.dispose()

@pytest_asyncio.fixture(name="session")
async def session_fixture(engine_fixture):
    async with engine_fixture.begin() as conn:
        await conn.run_sync(SQLModel.metadata.create_all)

    async_session = sessionmaker(
        engine_fixture, class_=AsyncSession, expire_on_commit=False
    )
    async with async_session() as session:
        yield session

    async with engine_fixture.begin() as conn:
        await conn.run_sync(SQLModel.metadata.drop_all)

@pytest_asyncio.fixture(name="client")
async def client_fixture(session: AsyncSession):
    def get_session_override():
        return session

    app.dependency_overrides[get_session] = get_session_override
    
    async with AsyncClient(transport=ASGITransport(app=app), base_url="http://test") as client:
        yield client
    
    app.dependency_overrides.clear()
